[OPIK-4568] [FE] Rename span-type-filter to camelCase spanTypeFilter
