+++ apps/opik-frontend/src/lib/spanTypeFilter.tsx
